+++ 몽실_김민정/week8/mission1/src/pages/Loading/Loading.js
@@ -1,4 +1,4 @@
-import { ReactComponent as LoadingSpinner } from "../assets/loadingSpinner.svg";
+import { ReactComponent as LoadingSpinner } from "../../assets/loadingSpinner.svg";
 import "./Loading.css";
 
 export const Loading = () => {
